python3/requirements.txt
# Upgrade pip: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple --upgrade pip
# Install with: pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt

#whell
wheel >= 0.30.0

# Excel
XlsxWriter >=1.0.2
openpyxl >= 2.5.1
xlutils >= 2.0.0

# MySQL
PyMySQL >= 0.7.11

# Requests
requests >= 2.18.4

# paramiko
paramiko >= 2.4.0

qcloudapi-sdk-python >= 2.0.11
elasticsearch >= 6.2.0
baidu-aip >= 2.2.